gcc warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@28294 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/common/crOpenGL/stub.c
@@ -467,7 +467,8 @@ GLboolean stubUpdateWindowVisibileRegions(WindowInfo *pWindow)
         pGLRects = crAlloc(4*cRects*sizeof(GLint));
         if (!pGLRects)
         {
-            crWarning("stubUpdateWindowVisibileRegions: failed to allocate %i bytes", 4*cRects*sizeof(GLint));
+            crWarning("stubUpdateWindowVisibileRegions: failed to allocate %lu bytes",
+                    (unsigned long)(4*cRects*sizeof(GLint)));
             return GL_FALSE;
         }
